+++ forAll/index.d.ts
@@ -1,2 +1,2 @@
-import { ReaxInstance, ReaxInstanceForAll } from '../types';
-export default function forAll(store: ReaxInstance): ReaxInstanceForAll;
+import forAll from './react';
+export default forAll;