+++ forFunctional/index.d.ts
@@ -1,2 +1,2 @@
-import { ReaxInstance, ReaxInstanceForFunctional } from '../types';
-export default function forFunctional(store: ReaxInstance): ReaxInstanceForFunctional;
+import forFunctional from './react';
+export default forFunctional;