+++ forFunctional/react/index.d.ts
@@ -1,0 +1,2 @@
+import { ReaxInstance } from '../../types';
+export default function forFunctional(store: ReaxInstance): import("../../types").ReaxInstanceForFunctional;