+++ types.d.ts
@@ -43,3 +43,8 @@ export interface ReaxInstanceForAll extends ReaxInstance {
     $$rebuildGetters: () => void;
     connectGettersToState: (ctx: React.Component, getter: string | string[]) => void;
 }
+export interface ReactLikeComponentClass {
+    setState: CallableFunction;
+    componentWillUnmount: CallableFunction;
+    state: object;
+}